requirements.txt
# Data validation and HTTP
pydantic
aiohttp
requests
streamlit

# Web framework and API
fastapi
uvicorn

pandas
scrapy
langchain-core
langchain
langchain-community
langchain-openai
rank_bm25
langchain_experimental
FlagEmbedding
qdrant-client
llama-index
llama-index-embeddings-huggingface
llama-index-vector-stores-qdrant
llama-index-retrievers-bm25
transformers
google-generativeai
llama-index-embeddings-huggingface
python-telegram-bot
langchain
chromadb
faiss-cpu
Flask
pdfplumber
gTTS
edge-tts
python-multipart
SpeechRecognition
noisereduce
scipy
pyaudio
pydub
pymupdf